Remove old tests, move fixtures
